Consolidate graph and security helpers
