+++ updaters/projectupdater.ts
@@ -7,7 +7,6 @@ import { GraphGroup } from "azure-devops-node-api/interfaces/GraphInterfaces";
 import { IProject, IProjectPermission } from "../interfaces/readers/configurationreader";
 import { IConsoleLogger } from "../interfaces/common/consolelogger";
 import { IDebugLogger } from "../interfaces/common/debuglogger";
-import { IGraphHelper } from "../interfaces/helpers/graphhelper";
 import { IHelper } from "../interfaces/common/helper";
 import { IProjectHelper } from "../interfaces/helpers/projecthelper";
 import { IProjectUpdater } from "../interfaces/updaters/projectupdater";
@@ -16,21 +15,19 @@ import { ISecurityHelper } from "../interfaces/helpers/securityhelper";
 export class ProjectUpdater implements IProjectUpdater {
 
     public projectHelper: IProjectHelper;
-    public graphHelper: IGraphHelper;
     public securityHelper: ISecurityHelper;
 
     private debugLogger: Debug.Debugger;
     private logger: IConsoleLogger;
     private helper: IHelper;
 
-    constructor(projectHelper: IProjectHelper, graphHelper: IGraphHelper, securityHelper: ISecurityHelper, debugLogger: IDebugLogger, consoleLogger: IConsoleLogger, helper: IHelper) {
+    constructor(projectHelper: IProjectHelper, securityHelper: ISecurityHelper, debugLogger: IDebugLogger, consoleLogger: IConsoleLogger, helper: IHelper) {
 
         this.debugLogger = debugLogger.create(this.constructor.name);
         this.logger = consoleLogger;
         this.helper = helper;
 
         this.projectHelper = projectHelper;
-        this.graphHelper = graphHelper;
         this.securityHelper = securityHelper;
 
     }
@@ -127,7 +124,7 @@ export class ProjectUpdater implements IProjectUpdater {
 
                 this.logger.log(`Updating <${groupName}> group members`);
 
-                await this.graphHelper.updateGroupMembers(group.members, targetGroup);
+                await this.securityHelper.updateGroupMembers(group.members, targetGroup);
 
             }
